Refactor Tenant
